Moved settings to separate folder
