+++ PasswordManager/App.xaml.cs
@@ -10,6 +10,7 @@ using PasswordManager.Clouds.Services;
 using PasswordManager.Helpers;
 using PasswordManager.Hotkeys;
 using PasswordManager.Services;
+using PasswordManager.Settings;
 using PasswordManager.ViewModels;
 using PasswordManager.Views;
 using System;
